点赞按钮动画添加

+++ TDjuwairen/Modules/Alive/AliveList/Cells/AliveListBottomTableViewCell.m
@@ -42,9 +42,9 @@
 - (void)setupAliveModel:(AliveListModel *)aliveModel {
     
     _cellModel = aliveModel;
-    [self.shareBtn setTitle:[NSString stringWithFormat:@"%ld", aliveModel.shareNum] forState:UIControlStateNormal];
-    [self.commentBtn setTitle:[NSString stringWithFormat:@"%ld", aliveModel.commentNum] forState:UIControlStateNormal];
-    [self.likeBtn setTitle:[NSString stringWithFormat:@"%ld", aliveModel.likeNum] forState:UIControlStateNormal];
+    [self.shareBtn setTitle:[NSString stringWithFormat:@"%ld", (long)aliveModel.shareNum] forState:UIControlStateNormal];
+    [self.commentBtn setTitle:[NSString stringWithFormat:@"%ld", (long)aliveModel.commentNum] forState:UIControlStateNormal];
+    [self.likeBtn setTitle:[NSString stringWithFormat:@"%ld", (long)aliveModel.likeNum] forState:UIControlStateNormal];
     self.likeBtn.selected = aliveModel.isLike;
 }
 @end
